+++ packages/react/src/index.ts
@@ -1,3 +1,4 @@
 export { ConsentBanner, type ConsentBannerProps } from './banner.js'
+export { ConsentNotice, type ConsentNoticeProps } from './notice.js'
 export { ConsentProvider, type ConsentProviderProps } from './provider.js'
 export { type CategoryApi, type ConsentApi, useConsent } from './use-consent.js'